+++ platform/go-server/internal/models/material.go
@@ -5,8 +5,8 @@ import (
 )
 
 type Material struct {
-	ID       int     `gorm:"primaryKey;autoIncrement" json:"id"`
-	CorpID   int     `gorm:"column:corp_id;default:0" json:"corp_id"`
+	ID       int64   `gorm:"primaryKey;autoIncrement" json:"id"`
+	CorpID   int64   `gorm:"column:corp_id;default:0" json:"corp_id"`
 	Name     string  `gorm:"column:name" json:"name"`
 	Spec     string  `gorm:"column:spec" json:"spec,omitempty"`
 	Quantity float64 `gorm:"column:quantity" json:"quantity"`
@@ -16,9 +16,9 @@ type Material struct {
 func (Material) TableName() string { return "materials" }
 
 type Remnant struct {
-	ID           int       `gorm:"primaryKey;autoIncrement" json:"id"`
-	CorpID       int       `gorm:"column:corp_id;default:0" json:"corp_id"`
-	MaterialID   int       `gorm:"column:material_id;index" json:"material_id"`
+	ID           int64     `gorm:"primaryKey;autoIncrement" json:"id"`
+	CorpID       int64     `gorm:"column:corp_id;default:0" json:"corp_id"`
+	MaterialID   int64     `gorm:"column:material_id;index" json:"material_id"`
 	MaterialName string    `gorm:"-" json:"material_name,omitempty"`
 	Dimensions   string    `gorm:"column:dimensions" json:"dimensions"`
 	PhotoData    string    `gorm:"column:photo_data" json:"photo_data,omitempty"`
